Render an hp bar

+++ Gameplay/Combat/IDamageablePlayer.cs
@@ -4,5 +4,5 @@ namespace Gameplay.Combat;
 
 internal interface IDamageablePlayer : ICircleCollider
 {
-    public float Health { get; set; }
+    public int Health { get; set; }
 }
